requirements.txt
# PDF OCR and Vision Analysis Pipeline

# Core dependencies
Pillow>=10.0.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0
customtkinter>=5.2.0
rapidocr-onnxruntime>=1.3.0
pywin32>=306; sys_platform == 'win32'

# SiliconFlow API
requests>=2.28.0

# Packaging
pyinstaller>=6.0.0
